lint, replace type with wallet-core
